all Them gio hàng

+++ DAL/Models/SanPham.cs
@@ -12,7 +12,7 @@ namespace DAL.Models
 
         public int Id { get; set; }
         public string Ten { get; set; } = null!;
-        public decimal? Gia { get; set; }
+        public int? Gia { get; set; }
         public int? Soluong { get; set; }
         public string? Mota { get; set; }
         public int? Trangthai { get; set; }
